Fix error with dupligroups
The functionality in 2.79 was covered by dupligroups and in 2.80 these became instance collections. A collection instance can only be made from an Empty object type. Haven't tested the functionality thoroughly, but it doesn't throw an error anymore.

There's still a matrix multiplication bug happening to be able to check objects that are not at 0, 0, 0 location.