Improve material and texture handling
Fixes errors when trying to export emission shader materials.

Exporting textures is tailored to what OpenMW requires which only needs a single "DiffuseMap" per material. Specular and normal maps are used automatically by OpenMW when they have a correct name and we don't need other maps.

Thus we only export a single texture that's connected to "Base Color" socket of the shader node in Blender's material. Logic is left in place to allow other texture types such as specular, normal maps etc. if there's ever a need in the future.