Improve collision shape workflow in Blender
Specific to OpenMW, a mesh needs to be a child of a node called "collision" or "Collision". This mesh will then not be rendered, but used for physics collisions.

Since objects in Blender need unique names, it makes for a nicer workflow to detect an assigned collision modifier and change the name that OpenMW recognizes.