Re-enable applying modifiers to meshes
Not only did the to_mesh() function change between Blender versions, but it's required that the object holding the mesh data is touched by the evaluated  dependency graph. Once this is done at both places, we get the result of applied modifiers.

When only the to_mesh() function was changed, modified mesh result wasn't produced.